🔥 Delete: 필요없는 더미데이터 삭제
