--- src/mocks/userData.ts
@@ -1,10 +1,0 @@
-import type { UserItem } from '@/pages/profile/account/types/account.type';
-
-export const userData: UserItem = {
-  userId: 1,
-  nickname: 'speakOnUser',
-  profileImgUrl: '',
-  name: '홍길동',
-  email: 'example@naver.com',
-  socialType: 'GOOGLE',
-};
