+++ src/pages/profile/payments/types/payments.type.ts
@@ -1,0 +1,9 @@
+/* 구독 정보 */
+export type PaymentMethod = '토스페이';
+
+export interface PaymentsItem {
+  name: string;
+  explain: string;
+  price: number;
+  paymentMethod: PaymentMethod;
+}
